requirements.txt
Django
djangorestframework
djongo
numpy
pymongo
scikit_learn
sentence_transformers
spacy
supabase
openai
tenacity
django-cors-headers
